MY AI/requirements.txt
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Session==0.5.0
cloudinary==1.36.0
python-dotenv==1.0.0
